requirements.txt
Appium-Python-Client
colorama
tqdm
requests